* Model :     - removed obsolete detection data   * Kernel :     - removed error messages for getting paramterers with default value

+++ tronc/sources/kernel/implementation/parameters.cpp
@@ -97,7 +97,7 @@ namespace ProjetUnivers
               break ;
             default:
               // try numeric
-              numeric_value = atof(value.c_str()) ;
+              numeric_value = (float)atof(value.c_str()) ;
               m_parameters[section][name] = numeric_value ;
             }
